requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2021.10.8
charset-normalizer==3.4.2
distlib==0.3.4
exceptiongroup==1.3.0
filelock==3.4.2
google-auth==2.40.3
google-genai==1.25.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pipenv==2022.1.8
platformdirs==2.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
requests==2.32.4
rsa==4.9.1
six==1.16.0
sniffio==1.3.1
tenacity==8.5.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
virtualenv==20.13.0
virtualenv-clone==0.5.7
websockets==15.0.1
